src/main/java/de/hpi/fgis/json/DBObjectStreamIterator.java
package de.hpi.fgis.json;

import java.io.BufferedReader;
import java.io.IOException;
import java.util.Iterator;
import java.util.NoSuchElementException;

import com.mongodb.DBObject;
import com.mongodb.util.JSON;

/**
 * utility to read a stream of tweet objects and returns single objects (all objects have to be split by newline(s))
 * @author tongr
 *
 */
public class DBObjectStreamIterator implements Iterator<DBObject> {
	private BufferedReader reader;
	private DBObject next;
	private final boolean fixNewlines;
	
	public DBObjectStreamIterator(BufferedReader reader, boolean fixNewlines) {
		this.reader = reader;
		this.fixNewlines = fixNewlines;
		
		setNext();
	}
	private void setNext() {
		next = null;
		// do all the magic
		if(reader!=null) {
			StringBuilder objTxt = new StringBuilder();
			String line;
			
			do {
				try {
					line = reader.readLine();
				} catch (IOException e) {
					line = null;
					reader = null;
				}
				// ignore "undefined" lines
				if(line!=null && !line.trim().equalsIgnoreCase("undefined")) {
					objTxt.append(line).append('\n');
					
					try {
						next = (DBObject) JSON.parse(objTxt.toString());
					} catch (RuntimeException e) {
						// unable to parse JSON object
						if(!fixNewlines) {
							// ignore the last line and continue with the next one (starting with a new object)
							objTxt.setLength(0);
						}
						// otherwise: nothing to do -> try to parse an object including the next line
					}
				}
			} while (next==null && line!=null);
		}
	}

	@Override
	public DBObject next() {
		DBObject current;
		synchronized (this) {
			current = next;
			if(current==null) {
				throw new NoSuchElementException();
			}
			
			setNext();	
		}
		return current;
	}

	@Override
	public boolean hasNext() {
		return next!=null;
	}

	@Override
	public void remove() {
		throw new IllegalStateException();
	}
	
}
